Move all primitive props to parent component
