+++ packages/react/src/components/AccountSettings/ChangePassword/defaultComponents.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { translate } from '@aws-amplify/ui';
 import { Alert, Button, PasswordField } from '../../../primitives';
 import {
   AccountSettingsError,
@@ -7,43 +6,19 @@ import {
   AccountSettingsSubmitButton,
 } from '../types';
 
+/** ChangePassword subcomponents */
+// TODO: enable component override
 export const DefaultCurrentPassword: AccountSettingsPasswordField = (props) => {
-  // TODO: add AccountSettingsTextUtil to collect these strings
-  const label = translate('Current Password');
-
-  return (
-    <PasswordField
-      label={label}
-      name="currentPassword"
-      autoComplete="current-password"
-      isRequired
-      {...props}
-    />
-  );
+  return <PasswordField {...props} />;
 };
 
 export const DefaultNewPassword: AccountSettingsPasswordField = (props) => {
-  const label = translate('New Password');
-
-  return (
-    <PasswordField
-      label={label}
-      name="newPassword"
-      autoComplete="new-password"
-      isRequired
-      {...props}
-    />
-  );
+  return <PasswordField {...props} />;
 };
 
 export const DefaultSubmitButton: AccountSettingsSubmitButton = (props) => {
-  const updatePasswordText = translate('Update password');
-
-  return (
-    <Button type="submit" {...props}>
-      {updatePasswordText}
-    </Button>
-  );
+  const { children, ...rest } = props;
+  return <Button {...rest}>{children}</Button>;
 };
 
 export const DefaultError: AccountSettingsError = (props) => {
